Move grid list and searching list from MainView to DirManipulation object
